29 april update

+++ manage_admin_settings/views.py
@@ -58,32 +58,40 @@ class EmailSettings(generic.TemplateView):
 
 
 	def get(self, request,*args, **kwargs):
-		email_setting_form = EmailSettingForm
+		if request.user.is_superuser:
+			email_setting_form = EmailSettingForm
 
-		# Check email setting already exist or not.
-		if Email_Setting.objects.filter(status=True).exists():
-			email_value = get_object_or_404(Email_Setting, status=True)
-			email_setting_form = EmailSettingForm(instance=email_value)
-	
-		return render(request, self.template_name,{'email_setting_form':email_setting_form,"username":request.user.username})
-
-	def post(self, request):
-		# Post function for add or update email setting.
-		if EmailSettingForm(request.POST):
-			email_setting_form = EmailSettingForm(request.POST)
 			# Check email setting already exist or not.
 			if Email_Setting.objects.filter(status=True).exists():
 				email_value = get_object_or_404(Email_Setting, status=True)
-				email_setting_form = EmailSettingForm(request.POST, instance=email_value)
-
-			if email_setting_form.is_valid():
-				data = email_setting_form.save(commit=False)
-				data.updated_dt = datetime.now()
-				data.save()
-			messages.success(request,'Email succesfully updated')
+				email_setting_form = EmailSettingForm(instance=email_value)
+		
+			return render(request, self.template_name,{'email_setting_form':email_setting_form,"username":request.user.username})
 		else:
-			messages.error(request,'Email can not be updated')
-		return HttpResponseRedirect(reverse('manage_admin_settings:EmailSettings'))
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
+
+	def post(self, request):
+		if request.user.is_superuser:
+			# Post function for add or update email setting.
+			if EmailSettingForm(request.POST):
+				email_setting_form = EmailSettingForm(request.POST)
+				# Check email setting already exist or not.
+				if Email_Setting.objects.filter(status=True).exists():
+					email_value = get_object_or_404(Email_Setting, status=True)
+					email_setting_form = EmailSettingForm(request.POST, instance=email_value)
+
+				if email_setting_form.is_valid():
+					data = email_setting_form.save(commit=False)
+					data.updated_dt = datetime.now()
+					data.save()
+				messages.success(request,'Email succesfully updated')
+			else:
+				messages.error(request,'Email can not be updated')
+			return HttpResponseRedirect(reverse('manage_admin_settings:EmailSettings'))
+		else:
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
 
 		
 @method_decorator(login_required, name="dispatch")
@@ -92,19 +100,23 @@ class ManageEnquiries(generic.TemplateView):
 
 
 	def get(self, request,*args, **kwargs):
-		from_date=None
-		to_date=None
-		if 'from_date' in request.GET and 'to_date' in request.GET :
-			from_date = request.GET['from_date']
-			to_date = request.GET['to_date']
-			
-			from_date_format = datetime.strptime(from_date,"%m/%d/%Y")
-			to_date_format = datetime.strptime(to_date,"%m/%d/%Y")
-			
-			get_enquiries=ContactUs.objects.filter(created_date__range=[from_date_format, to_date_format])
+		if request.user.is_superuser:
+			from_date=None
+			to_date=None
+			if 'from_date' in request.GET and 'to_date' in request.GET :
+				from_date = request.GET['from_date']
+				to_date = request.GET['to_date']
+				
+				from_date_format = datetime.strptime(from_date,"%m/%d/%Y")
+				to_date_format = datetime.strptime(to_date,"%m/%d/%Y")
+				
+				get_enquiries=ContactUs.objects.filter(created_date__range=[from_date_format, to_date_format])
+			else:
+				get_enquiries=ContactUs.objects.all()
+			return render(request, self.template_name,{'get_enquiries':get_enquiries,'from_date':from_date,'to_date':to_date})
 		else:
-			get_enquiries=ContactUs.objects.all()
-		return render(request, self.template_name,{'get_enquiries':get_enquiries,'from_date':from_date,'to_date':to_date})
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
 
 
 
@@ -115,44 +127,52 @@ class EnquiryReply(generic.TemplateView):
 
 
 	def get(self, request,id,*args, **kwargs):
-		
-		get_instance=get_object_or_404(ContactUs,id=id)
-		return render(request, self.template_name,{'get_instance':get_instance})
+		if request.user.is_superuser:
+			
+			get_instance=get_object_or_404(ContactUs,id=id)
+			return render(request, self.template_name,{'get_instance':get_instance})
+		else:
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
 
 	def post(self, request,id,*args, **kwargs):
-		subject=request.POST['subject']
-		print(subject)
-		comment=request.POST['description']
-		get_instance=get_object_or_404(ContactUs,id=id)
+		if request.user.is_superuser:
+			subject=request.POST['subject']
+			# print(subject)
+			comment=request.POST['description']
+			get_instance=get_object_or_404(ContactUs,id=id)
 
-		if User.objects.filter(email=get_instance.email).exists():
-		   
-			
-			data_content = {"username": get_instance.name,
-							"comment": comment}
-			email_content = 'admin/email-pages/reply-email.html'
+			if User.objects.filter(email=get_instance.email).exists():
+			   
+				
+				data_content = {"username": get_instance.name,
+								"comment": comment}
+				email_content = 'admin/email-pages/reply-email.html'
 
-			email_template = get_template(email_content).render(data_content)
-			reciver_email = get_instance.email
+				email_template = get_template(email_content).render(data_content)
+				reciver_email = get_instance.email
 
-			Subject = subject
-			if Email_Setting.objects.filter(status=True).exists():
-				email_data = Email_Setting.objects.filter(status=True)
-				for data in email_data:
-					EMAIL_HOST = data.smtp_host
-					EMAIL_PORT = data.smtp_port
-					EMAIL_HOST_USER = data.smtp_username
-					EMAIL_HOST_PASSWORD = data.smtp_password
-			email_msg = EmailMessage(Subject, email_template, EMAIL_HOST_USER, [reciver_email],
-									 reply_to=[EMAIL_HOST_USER])
-			email_msg.content_subtype = 'html'
-			email_msg.send(fail_silently=False)
-			get_instance.reply_status=True
-			get_instance.save()
-			messages.success(request, "successfully replied to "+get_instance.name+"")
+				Subject = subject
+				if Email_Setting.objects.filter(status=True).exists():
+					email_data = Email_Setting.objects.filter(status=True)
+					for data in email_data:
+						EMAIL_HOST = data.smtp_host
+						EMAIL_PORT = data.smtp_port
+						EMAIL_HOST_USER = data.smtp_username
+						EMAIL_HOST_PASSWORD = data.smtp_password
+				email_msg = EmailMessage(Subject, email_template, EMAIL_HOST_USER, [reciver_email],
+										 reply_to=[EMAIL_HOST_USER])
+				email_msg.content_subtype = 'html'
+				email_msg.send(fail_silently=False)
+				get_instance.reply_status=True
+				get_instance.save()
+				messages.success(request, "successfully replied to "+get_instance.name+"")
+			else:
+				messages.error(request, "something going wrong please try again")
+			return HttpResponseRedirect(reverse('manage_admin_settings:ManageEnquiries'))
 		else:
-			messages.error(request, "something going wrong please try again")
-		return HttpResponseRedirect(reverse('manage_admin_settings:ManageEnquiries'))
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
 		   
 
 
@@ -161,11 +181,15 @@ class EnquiryReply(generic.TemplateView):
 class DeleteMultiEnquiries(generic.TemplateView):
 
 	def post(self, request,*args, **kwargs):
-		enquiries_array=request.POST.getlist('enquiries_array[]')
-		
-		for data in enquiries_array:
-			enquiry_instance = get_object_or_404(ContactUs, id=data)
+		if request.user.is_superuser:
+			enquiries_array=request.POST.getlist('enquiries_array[]')
 			
-			ContactUs.objects.filter(id=enquiry_instance.id).delete()
-		
-		return JsonResponse({'message': 'Deleted successfully.'})
+			for data in enquiries_array:
+				enquiry_instance = get_object_or_404(ContactUs, id=data)
+				
+				ContactUs.objects.filter(id=enquiry_instance.id).delete()
+			
+			return JsonResponse({'message': 'Deleted successfully.'})
+		else:
+			messages.error(request,"Please enter valid username and password")
+			return HttpResponseRedirect(reverse('admin_login'))
